fix comission equation

+++ src/main/java/com/bayzdelivery/model/Delivery.java
@@ -37,7 +37,7 @@ public class Delivery implements Serializable{
   Long price;
 
   @Column(name = "comission")
-  Long comission;
+  Double comission;
 
   @ManyToOne
   @JoinColumn(name = "delivery_man_id", referencedColumnName = "id")
@@ -83,11 +83,11 @@ public class Delivery implements Serializable{
     this.price = price;
   }
 
-  public Long getComission() {
+  public Double getComission() {
     return comission;
   }
 
-  public void setComission(Long comission) {
+  public void setComission(Double comission) {
     this.comission = comission;
   }
